+++ src/Resources/contao/dca/tl_content.php
@@ -22,8 +22,8 @@ $strTable = 'tl_content';
  * Add palette to tl_content.
  */
 
-$GLOBALS['TL_DCA'][$strTable]['palettes']['mobileDeList'] = '{type_legend},type,headline;{md_settings_legend},pdirVehicleFilterByAccount,pdir_md_hidePromotionBox,pdir_md_listTemplate,pdir_md_itemTemplate,pdir_md_readerPage,pdir_md_listingPage;{md_filters_legend},pdirVehicleFilterFields,pdirVehicleFilterWhere,pdirVehicleFilterMaxItems,pdirVehicleFilterSort,pdir_md_hideFilters,pdir_open_filter,pdir_md_list_shuffle,pdir_md_priceSlider,pdir_md_powerSlider,pdir_md_mileageSlider,pdir_md_only_filter;{md_template_legend},pdir_md_promotion_corner_position,pdir_md_promotion_corner_color,pdir_md_promotion_corner_shadow,pdir_md_corner_position,pdir_md_corner_color,pdir_md_corner_shadow,pdir_md_removeModuleCss,pdir_md_removeModuleJs;{md_debug_legend},pdir_md_enableDebugMode,pdir_md_forceRefresh,pdir_md_cacheTime;{expert_legend:hide},cssID,space';
-$GLOBALS['TL_DCA'][$strTable]['palettes']['mobileDeReader'] = '{type_legend},type,headline;{md_settings_legend},pdirVehicleFilterByAccount,pdir_md_hidePromotionBox,pdir_md_corner_position,pdir_md_corner_color,pdir_md_corner_shadow,pdir_md_removeModuleCss,pdir_md_removeModuleJs;{md_debug_legend},pdir_md_enableDebugMode;{expert_legend:hide},cssID,space';
+$GLOBALS['TL_DCA'][$strTable]['palettes']['mobileDeList'] = '{type_legend},type,headline;{md_settings_legend},pdirVehicleFilterByAccount,pdir_md_hidePromotionBox,pdir_md_listTemplate,pdir_md_itemTemplate,pdir_md_readerPage,pdir_md_listingPage;{md_filters_legend},pdirVehicleFilterFields,pdirVehicleFilterWhere,pdirVehicleFilterMaxItems,pdirVehicleFilterSort,pdir_md_hideFilters,pdir_open_filter,pdir_md_list_shuffle,pdir_md_priceSlider,pdir_md_powerSlider,pdir_md_mileageSlider,pdir_md_only_filter;{md_template_legend},pdir_md_removeModuleCss,pdir_md_removeModuleJs;{md_debug_legend},pdir_md_enableDebugMode,pdir_md_forceRefresh,pdir_md_cacheTime;{expert_legend:hide},cssID,space';
+$GLOBALS['TL_DCA'][$strTable]['palettes']['mobileDeReader'] = '{type_legend},type,headline;{md_settings_legend},pdirVehicleFilterByAccount,pdir_md_removeModuleCss,pdir_md_removeModuleJs;{md_debug_legend},pdir_md_enableDebugMode;{expert_legend:hide},cssID,space';
 
 /*
  * Add fields to tl_content
@@ -108,80 +108,6 @@ $GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_list_shuffle'] = [
     'eval' => [
         'submitOnChange' => true,
         'tl_class' => 'w50',
-    ],
-    'sql' => "char(1) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_corner_color'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color'],
-    'exclude' => true,
-    'inputType' => 'select',
-    'options' => [
-        'white' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_white'],
-        'black' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_black'],
-        'grey' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_grey'],
-        'blue' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_blue'],
-        'green' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_green'],
-        'turquoise' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_turquoise'],
-        'purple' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_purple'],
-        'red' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_red'],
-        'orange' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_orange'],
-        'yellow' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_color_yellow'],
-    ],
-    'eval' => ['tl_class' => 'w50'],
-    'sql' => "varchar(64) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_promotion_corner_color'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_promotion_corner_color'],
-    'exclude' => true,
-    'inputType' => 'select',
-    'options' => &$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_corner_color']['options'],
-    'eval' => ['tl_class' => 'w50'],
-    'sql' => "varchar(64) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_corner_position'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_position'],
-    'exclude' => true,
-    'inputType' => 'select',
-    'options' => [
-        'top-left' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_position_top_left'],
-        'top-right' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_position_top_right'],
-        'bottom-left' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_position_bottom_left'],
-        'bottom-right' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_position_bottom_right'],
-    ],
-    'eval' => ['tl_class' => 'w50 clr'],
-    'sql' => "varchar(64) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_promotion_corner_position'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_promotion_corner_position'],
-    'exclude' => true,
-    'inputType' => 'select',
-    'options' => &$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_corner_position']['options'],
-    'eval' => ['tl_class' => 'w50 clr'],
-    'sql' => "varchar(64) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_promotion_corner_shadow'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_promotion_corner_shadow'],
-    'exclude' => true,
-    'inputType' => 'checkbox',
-    'eval' => [
-        'submitOnChange' => true,
-        'tl_class' => 'w50 m12',
-    ],
-    'sql' => "char(1) NOT NULL default ''",
-];
-
-$GLOBALS['TL_DCA'][$strTable]['fields']['pdir_md_corner_shadow'] = [
-    'label' => &$GLOBALS['TL_LANG'][$strTable]['pdir_md_corner_shadow'],
-    'exclude' => true,
-    'inputType' => 'checkbox',
-    'eval' => [
-        'submitOnChange' => true,
-        'tl_class' => 'm12 clr',
     ],
     'sql' => "char(1) NOT NULL default ''",
 ];